+++ src/pages/manager-ipr-draft/manager-ipr-draft.tsx
@@ -4,7 +4,6 @@ import React, { FC, ChangeEvent, useState } from 'react';
 import { Footer } from '../../entities/footer/footer';
 
 import Header from '../../shared/header-component/header';
-import NavBar from '../../entities/navbar/navbar';
 
 import { EmployeeInfoCard } from '../../entities/employee-info-card/employee-info-card';
 import { Button } from '@alfalab/core-components/button';
@@ -21,6 +20,7 @@ import avatar from '../../images/avatars/avatar_head-of-dept.png';
 import { Tasks } from '../../entities/tasks/tasks';
 import NavBarMini from '../../entities/navbar-mini/navbar-mini';
 import { Modal } from '../../entities/modal/modal';
+import { TasksOverview } from '../../entities/tasks-overview/tasks-overview';
 interface ManagerIprDraftProps {
 	statusText: string;
 	statusColor: StatusProps['color'];
@@ -45,222 +45,11 @@ export const ManagerIprDraft = ({
 	// 		avatar: avatar,
 	// 	},
 	// ];
-	const optionsGoal: OptionShape[] = [
-		{ key: 'Карьерный рост' },
-		{ key: 'Повышение грейда' },
-		{ key: 'Соответствие занимаемой должности' },
-		{ key: 'Развитие софт-скиллов' },
-		{ key: 'Развитие хард-скиллов' },
-		{ key: 'Смена специализации' },
-		{ key: 'Смена команды' },
-		{ key: 'Получение нового опыта' },
-	];
-
-	const optionsRole: OptionShape[] = [
-		{ key: 'Продакт-менеджер' },
-		{ key: 'Проджект-менеджер' },
-		{ key: 'Бизнес-аналитик' },
-		{ key: 'Системный аналитик' },
-		{ key: 'Дизайнер' },
-		{ key: 'QA-инженер' },
-		{ key: 'Фронтенд-разработчик' },
-		{ key: 'Бэкенд-разработчик' },
-		{ key: 'Мобильный разработчик' },
-		{ key: 'DevOps-инженер' },
-		{ key: 'Системный администратор' },
-		{ key: 'Дата-аналитик' },
-		{ key: 'Дата-сайентист' },
-		{ key: 'Руководитель' },
-		{ key: 'HR' },
-		{ key: 'Другое' },
-	];
-
-	const optionsCompetence: OptionShape[] = [
-		{ key: 'Адаптивность и стрессоустойчивость' },
-		{ key: 'Анализ и решение проблем' },
-		{ key: 'Английский язык' },
-		{ key: 'Аналитическое мышление' },
-		{ key: 'Ведение переговоров' },
-		{ key: 'Выявление паттернов или связей' },
-		{ key: 'Гибкость и готовность к изменениям' },
-		{ key: 'Делегирование' },
-		{ key: 'Исполнительская дисциплина' },
-		{ key: 'Клиентоориентированность' },
-		{ key: 'Креативное мышление' },
-		{ key: 'Критическое мышление' },
-		{ key: 'Лидерство' },
-		{ key: 'Навык деловой переписки' },
-		{ key: 'Наставничество и менторинг' },
-		{ key: 'Ориентация на результат' },
-		{ key: 'Ответственность' },
-		{ key: 'Планирование и организация' },
-		{ key: 'Поиск информации и ресурсов для решения задач' },
-		{ key: 'Понимание бизнеса и структуры организации' },
-		{ key: 'Построение эффективных процессов' },
-		{ key: 'Принятие управленческих решений' },
-		{ key: 'Публичные выступления' },
-		{ key: 'Работа в команде' },
-		{ key: 'Работа в режиме неопределенности' },
-		{ key: 'Работоспособность и личная эффективность' },
-		{ key: 'Самоанализ и саморефлексия' },
-		{ key: 'Самомотивация' },
-		{ key: 'Самосовершенствование (самообучение)' },
-		{ key: 'Самостоятельность (самоорганизация)' },
-		{ key: 'Системное мышление' },
-		{ key: 'Следование стандартам деятельности' },
-		{ key: 'Способности к анализу и прогнозированию результатов' },
-		{ key: 'Стратегическое мышление' },
-		{ key: 'Умение обрабатывать полученную информацию' },
-		{ key: 'Умение оценивать сроки выполнения задач' },
-		{ key: 'Умение проводить исследования' },
-		{ key: 'Умение слушать' },
-		{ key: 'Управление данными' },
-		{ key: 'Управление изменениям' },
-		{ key: 'Управление конфликтами' },
-		{ key: 'Управление людьми' },
-		{ key: 'Управление проектами' },
-		{ key: 'Целеустремленность' },
-		{ key: 'Эмоциональный интеллект' },
-		{ key: 'Эмпатия' },
-		{ key: 'Другое' },
-	];
-
-	const optionsMentor: OptionShape[] = [
-		{ key: 'Иванова Наталья Дмитриевна' },
-		{ key: 'Петрова Наталья Дмитриевна' },
-		{ key: 'Сидорова Наталья Дмитриевна' },
-	];
-
-	const [multiple, setMultiple] = useState(true);
-	const [shownChevron, setShownChevron] = useState(true);
-
-	const [valueGoal, setValueGoal] = useState<string>('');
-	const [valueRole, setValueRole] = useState<string>('');
-	const [valueMentor, setValueMentor] = useState<string>('');
-	const [valueCompetence, setValueCompetence] = useState<string>('');
-	const [valueStartDate, setStartDate] = useState<string>('');
-	const [valueEndDate, setEndDate] = useState<string>('');
 
 	const [modalOpen, setModalOpen] = useState(false);
 
 	const onModalOpen = () => {
 		setModalOpen(!modalOpen);
-	};
-	const matchOption = (option: OptionShape, inputValue: string): boolean =>
-		option.key.toLowerCase().includes((inputValue || '').toLowerCase());
-
-	// Обработка импутов
-	const handleInputGoal = (
-		event: ChangeEvent<HTMLInputElement> | null,
-		{ value }: { value: string }
-	) => {
-		setValueGoal(value);
-	};
-
-	const handleInputRole = (
-		event: ChangeEvent<HTMLInputElement> | null,
-		{ value }: { value: string }
-	) => {
-		setValueRole(value);
-	};
-
-	const handleInputMentor = (
-		event: ChangeEvent<HTMLInputElement> | null,
-		{ value }: { value: string }
-	) => {
-		setValueMentor(value);
-	};
-	// Обработка изменения импутов
-
-	const handleChangeGoal = ({ selected }: { selected: OptionShape | null }) => {
-		setValueGoal(selected ? selected.key : '');
-	};
-
-	const handleChangeRole = ({ selected }: { selected: OptionShape | null }) => {
-		setValueRole(selected ? selected.key : '');
-	};
-	const handleChangeMentor = ({
-		selected,
-	}: {
-		selected: OptionShape | null;
-	}) => {
-		setValueMentor(selected ? selected.key : '');
-	};
-
-	const handleChangeStartDate = (event: any, { value }: { value: string }) => {
-		setStartDate(value);
-	};
-	const handleChangeEndDate = (event: any, { value }: { value: string }) => {
-		setEndDate(value);
-	};
-	// Обработка фильтры поиска значения
-
-	const getFilteredGoals = (): OptionShape[] => {
-		return optionsGoal.some(({ key }) => key === valueGoal)
-			? optionsGoal
-			: optionsGoal.filter((option) => matchOption(option, valueGoal));
-	};
-
-	const getFilteredRoles = (): OptionShape[] => {
-		return optionsRole.some(({ key }) => key === valueRole)
-			? optionsRole
-			: optionsRole.filter((option) => matchOption(option, valueRole));
-	};
-	const getFilteredMentor = (): OptionShape[] => {
-		return optionsMentor.some(({ key }) => key === valueMentor)
-			? optionsMentor
-			: optionsMentor.filter((option) => matchOption(option, valueMentor));
-	};
-
-	const handleInputCompetence = (
-		event: ChangeEvent<HTMLInputElement> | null,
-		{ value }: { value: string }
-	) => {
-		setValueCompetence(value);
-	};
-	const inputValues: string[] = valueCompetence.replace(/ /g, '').split(',');
-	const selectedOptions: OptionShape[] = optionsCompetence.filter((option) =>
-		inputValues.includes(option.key.trim())
-	);
-
-	const selected: string[] | OptionShape = multiple
-		? selectedOptions.map((option) => option.key)
-		: optionsCompetence.find((o) => o.key === inputValues[0]) || [];
-
-	const tagValues = valueCompetence.trim().split(',');
-
-	const handleChangeCompetence = ({
-		selected,
-		selectedMultiple,
-	}: {
-		selected: OptionShape | null;
-		selectedMultiple: OptionShape[] | null;
-	}): void => {
-		if (multiple) {
-			const value = selectedMultiple?.length
-				? selectedMultiple.map((option) => option.key).join(', ') + ', '
-				: '';
-			setValueCompetence(value);
-			return;
-		}
-		setValueCompetence(selected ? selected.key : '');
-	};
-
-	const getFilteredOptionsCompetence = (): OptionShape[] => {
-		if (multiple) {
-			return optionsCompetence.filter((option) => {
-				return (
-					selectedOptions.includes(option) ||
-					matchOption(option, inputValues[inputValues.length - 1])
-				);
-			});
-		}
-
-		return optionsCompetence.some(({ key }) => key === valueCompetence)
-			? optionsCompetence
-			: optionsCompetence.filter((option) =>
-					matchOption(option, valueCompetence)
-				);
 	};
 
 	return (
@@ -297,7 +86,6 @@ export const ManagerIprDraft = ({
 						<Button view="primary" size="m" className={styles.buttonSend}>
 							Отправить в работу
 						</Button>
-
 						<button
 							onClick={onModalOpen}
 							style={{
@@ -310,201 +98,7 @@ export const ManagerIprDraft = ({
 						</button>
 					</div>
 					<form className={styles2.form}>
-						<fieldset className={styles2.blockWrapper}>
-							<legend className={styles2.blockTitle}>Общее описание</legend>
-							<div className={styles2.formBlock}>
-								<div className={styles2.formRow}>
-									<div style={{ width: 468 }}>
-										<InputAutocomplete
-											name={'goal'}
-											block={true}
-											closeOnSelect={true}
-											className="inputGoal"
-											size="s"
-											options={getFilteredGoals()}
-											label="Цель *"
-											placeholder="Начните вводить название"
-											onChange={handleChangeGoal}
-											onInput={handleInputGoal}
-											Arrow={shownChevron ? Arrow : undefined}
-											value={valueGoal}
-											allowUnselect={true}
-											showEmptyOptionsList={true}
-											inputProps={{
-												onClear: () => setValueGoal(''),
-												clear: true,
-											}}
-										></InputAutocomplete>
-									</div>
-									<div style={{ width: 468 }}>
-										<InputAutocomplete
-											name={'role'}
-											block={true}
-											closeOnSelect={true}
-											className="inputRole"
-											size="s"
-											options={getFilteredRoles()}
-											label="Специализация *"
-											placeholder="Начните вводить название"
-											onChange={handleChangeRole}
-											onInput={handleInputRole}
-											Arrow={shownChevron ? Arrow : undefined}
-											value={valueRole}
-											allowUnselect={true}
-											inputProps={{
-												onClear: () => setValueRole(''),
-												clear: true,
-											}}
-										></InputAutocomplete>
-									</div>
-								</div>
-								<div style={{ width: 720 }}>
-									<InputAutocomplete
-										name={'competence'}
-										value={valueCompetence}
-										block={true}
-										multiple={multiple}
-										allowUnselect={true}
-										closeOnSelect={true}
-										onChange={handleChangeCompetence}
-										onInput={handleInputCompetence}
-										options={getFilteredOptionsCompetence()}
-										Arrow={shownChevron ? Arrow : undefined}
-										inputProps={{
-											onClear: () => setValueCompetence(''),
-											clear: true,
-										}}
-										showEmptyOptionsList={true}
-										className={styles.inputCompetence}
-										size="s"
-										label="Компетенция *"
-										placeholder="Начните вводить название"
-									></InputAutocomplete>
-								</div>
-								<div className={styles2.formRowTag}>
-									{valueCompetence.length > 0
-										? tagValues.map((value: string, key: number) => {
-												return (
-													<div
-														key={value.length + 1}
-														style={{ maxWidth: '319' }}
-													>
-														<FilterTag
-															showClear={true}
-															size="xxs"
-															shape="rounded"
-															view="filled"
-															checked={true}
-															onClear={() => setValueCompetence('')}
-														>
-															{value}
-														</FilterTag>
-													</div>
-												);
-											})
-										: ''}
-								</div>
-								<div className={styles2.formRow}>
-									<div style={{ width: 468 }}>
-										<InputAutocomplete
-											name={'mentor'}
-											block={true}
-											closeOnSelect={true}
-											className={styles.inputMentor}
-											size="s"
-											options={getFilteredMentor()}
-											label="Ментор"
-											placeholder="Начните вводить название"
-											onChange={handleChangeMentor}
-											onInput={handleInputMentor}
-											Arrow={shownChevron ? Arrow : undefined}
-											value={valueMentor}
-											allowUnselect={true}
-											inputProps={{
-												onClear: () => setValueMentor(''),
-												clear: true,
-											}}
-										></InputAutocomplete>
-									</div>
-
-									<div style={{ width: 222 }}>
-										<UniversalDateInput
-											block={true}
-											view="date"
-											label="Дата создания"
-											size="s"
-											value={valueStartDate}
-											onChange={handleChangeStartDate}
-											picker={true}
-											Calendar={CalendarDesktop}
-											calendarProps={{
-												selectorView: 'month-only',
-											}}
-											clear={true}
-											onClear={(e) => {
-												e.stopPropagation();
-												setStartDate('');
-											}}
-										/>
-									</div>
-									<div style={{ width: 222 }}>
-										<UniversalDateInput
-											block={true}
-											view="date"
-											label="Дата завершения"
-											size="s"
-											value={valueEndDate}
-											onChange={handleChangeEndDate}
-											picker={true}
-											Calendar={CalendarDesktop}
-											calendarProps={{
-												selectorView: 'month-only',
-											}}
-											clear={true}
-											onClear={(e) => {
-												e.stopPropagation();
-												setEndDate('');
-											}}
-										/>
-									</div>
-									<div
-										style={{
-											width: 960,
-										}}
-									>
-										<Textarea
-											fieldClassName={styles2.textClass}
-											maxHeight={91}
-											label="Описание"
-											labelView="inner"
-											size="m"
-											block={true}
-											maxLength={96}
-											showCounter={true}
-											autosize={true}
-										/>
-									</div>
-									<div
-										style={{
-											width: 960,
-										}}
-									>
-										<Textarea
-											fieldClassName={styles2.textClass}
-											maxHeight={91}
-											label="Комментарий (виден только вам)"
-											labelView="inner"
-											size="m"
-											block={true}
-											maxLength={96}
-											showCounter={true}
-											autosize={true}
-										/>
-									</div>
-								</div>
-							</div>
-						</fieldset>
-
+						<TasksOverview isExecutive={false}></TasksOverview>
 						<fieldset className={styles2.blockWrapper}>
 							<legend className={styles2.blockTitle}>Задачи</legend>
 							<Tasks isEmployee={false} />
